+++ src/App.js
@@ -1,10 +1,10 @@
 import { Grid } from '@material-ui/core';
 import { useDispatch, useSelector } from 'react-redux';
 import { ThemeProvider,createMuiTheme } from '@material-ui/core/styles';
-import Appbar from './ComponentsOne/Appbar';
-import CardItem from './ComponentsOne/CardItem';
-import { useEffect, useState } from 'react';
-import { replaceArray } from './ComponentsOne/redux/taskNamesArraySlice';
+import Appbar from './Components/Appbar';
+import CardItem from './Components/CardItem';
+import { useEffect } from 'react';
+import { replaceArray } from './Components/redux/taskNamesArraySlice';
 
 const theme = createMuiTheme({});
 
@@ -17,17 +17,16 @@ const App = () => {
       let obj = JSON.parse(arr)
       dispatch(replaceArray(obj))
     }
-  },[])
+  },[dispatch])
 
 
   return (
     <ThemeProvider theme={theme}>
       <Appbar/>
       <Grid container style={{flexGrow:1,alignContent:'center',padding:30}} spacing={2}>
-        {console.log('rend'+taskNames),
+        {
           taskNames && taskNames.map((task,index)=>(
-            <Grid key={index} item xm={12} sm={6} md={4}>
-              {console.log(index)}
+            <Grid key={index} item xm={12} sm={6} md={3}>
               <CardItem index={index}/>
             </Grid>
           ))
