+++ src/Components/ModalAddTask.jsx
@@ -5,8 +5,8 @@ import {  Button, TextareaAutosize, TextField } from '@material-ui/core';
 import { useDispatch, useSelector } from 'react-redux';
 import { setDescription } from './redux/TaskDescriptionSlice';
 import { setName } from './redux/TaskNameSlice';
-import {  replaceArray, setTasksArray } from './redux/taskNamesArraySlice';
-import { updateItem } from './redux/OpenSlice';
+import {  setTasksArray } from './redux/taskNamesArraySlice';
+import { toggle } from './redux/OpenSlice';
 
 const useStyles = makeStyles((theme) => ({
   modal: {
@@ -24,22 +24,20 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-export default function ModalUpadteTask({index}) {
+export default function ModalUI() {
   const classes = useStyles();
-  const edit = useSelector((state)=>state.open.edit)
+  const open = useSelector((state)=>state.open.open)
   const tName = useSelector((state)=>state.taskName.name)
-
   const tDescription = useSelector((state)=>state.taskDescription.description)
   const dispatch = useDispatch();
   const taskNames = useSelector((state) => state.taskNamesArray.tasks)
-
+  const edit = useSelector((state)=>state.open.edit)
 
   const handleSubmit = () =>{
-      dispatch(updateItem());
+      dispatch(toggle());
       let obj = {'name':tName,'description':tDescription,'date':new Date().toLocaleString()}
       let tempArray = [...taskNames,obj];
-      tempArray.splice(index,1)
-      dispatch(replaceArray(tempArray))
+      dispatch(setTasksArray(obj))
       localStorage.setItem('tasks',JSON.stringify(tempArray))
       dispatch(setName(''))
       dispatch(setDescription(''))
@@ -52,14 +50,14 @@ export default function ModalUpadteTask({index}) {
         aria-labelledby="transition-modal-title"
         aria-describedby="transition-modal-description"
         className={classes.modal}
-        open={edit}
+        open={open}
       >
           <div className={classes.paper}>
-            <h2 id="transition-modal-title">Update Task</h2>
+            <h2 id="transition-modal-title">{edit?'Update Task':'Add Task'}</h2>
             <form noValidate autoComplete="off">
-                <TextField id="outlined-basic" label="Task Name" variant="outlined" onChange={(e)=>dispatch(setName(e.target.value))} value={tName}/><br/>
+                <TextField id="outlined-basic" label="Task Name" variant="outlined" style={{width:'100%'}} onChange={(e)=>dispatch(setName(e.target.value))} value={tName}/><br/>
                 <TextareaAutosize onChange={(e)=>dispatch(setDescription(e.target.value))} value={tDescription} style={{marginTop:20,width:'100%',height:'100%',overflowY:'auto'}} rowsMin={5}  placeholder="Task Description" /> 
-                <Button style={{flexFlow:'row-reverse'}} onClick={handleSubmit} variant="contained" color="primary">Update</Button>
+                <Button style={{flexFlow:'row-reverse'}} onClick={handleSubmit} variant="contained" color="primary">Submit</Button>
             </form>
           </div>
       </Modal>
