+++ src/Components/redux/configurationStore.js
