+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
 import { Provider } from 'react-redux';
-import store from './ComponentsOne/redux/configurationStore';
+import store from './Components/redux/configurationStore';
 
 
 ReactDOM.render(
